+++ CHANGELOG.rst
@@ -6,11 +6,11 @@
 Change Log
 ::::::::::
 
-0.38.0 rc
-=========
+0.38.0
+======
 
 * Fixed a bug when building an assembly with Entity links.
-* TODO: Consider scene specs as regex strings.
+* First online documentation.
 
 0.37.0
 ======